rose build all passes for py37 and gfortran
