+++ run_configs/_rose_all/rose-suite.conf
@@ -50,11 +50,17 @@ source=$PWD/../lfric/lfric_common.py
 [file:bin/grab_lfric.py]
 source=$PWD/../lfric/grab_lfric.py
 
-[file:bin/mesh_tools.py]
-source=$PWD/../lfric/mesh_tools.py
-
 [file:bin/gungho.py]
 source=$PWD/../lfric/gungho.py
 
 [file:bin/atm.py]
 source=$PWD/../lfric/atm.py
+
+[file:bin/mesh_tools.py]
+source=$PWD/../lfric/mesh_tools.py
+
+[file:bin/mesh_tools_overrides]
+source=$PWD/../lfric/mesh_tools_overrides
+
+[file:bin/psyclone.cfg]
+source=$PWD/../lfric/psyclone.cfg
