+++ tests/unit_tests/steps/test_compile_fortran.py
@@ -8,7 +8,8 @@ import pytest
 from fab.build_config import BuildConfig
 from fab.constants import BUILD_TREES, OBJECT_FILES
 from fab.parse.fortran import AnalysedFortran
-from fab.steps.compile_fortran import CompileFortran, get_fortran_compiler, get_fortran_preprocessor, get_mod_hashes
+from fab.steps.compile_fortran import CompileFortran, get_fortran_compiler, get_mod_hashes
+from fab.steps.preprocess import get_fortran_preprocessor
 from fab.util import CompiledFile
